Fix off-by-one in apathmax() as the path is relative to "/"
1) Use size_t * instead of long *
2) Fallback to PATH_MAX instead of BUFSIZ
3) Header cleanup

+++ libutil/cp.c
@@ -28,7 +28,7 @@ cp(const char *s1, const char *s2, int depth)
 {
 	FILE *f1, *f2;
 	char *ns1, *ns2;
-	long size1, size2;
+	size_t size1, size2;
 	struct dirent *d;
 	struct stat st;
 	struct utimbuf ut;
